Add working version of simple code - pre cache
